Android - move mupdf SDK and open source code to a single folder
This is the third of three parts in preparing an open source
muPDF SDK. In parts 1 and 2 we changed the code architecture making
it possible to separate out things that could be made open source.
In this third part, we move the SDK and its open source code into
a separate folder.

A new folder called mupdf-sdk exists at the top, and contains
the sample app (mupdf-test), and editor, sodk_resourcessolib and wheel modules. A developer should be able to build
the SDK from everything that's in mupdf-sdk.

The smart-office-nui project has been modified so it uses open
source code and resources found under mupdf-sdk, along with any
SmartOffice-only things that are needed to build the retail flavors,
as well as the SO SDK.

ATS build scripts for MuPDF SDK and SO SDK are modified to
account for new locations.

<log>Android - move mupdf SDK and open source code to a single folder</log>