Android NUI: Remove Clipboard Change Listener
As of Android 10 we can no longer rely on the clipboard change
handler getting called we the app is backgrounded i.e. if the
user switches to another app to copy text.

Android 10 Release Notes
------------------------

"Restrict app clipboard access

In Android 10, clipboard access has changed so that clipboard content
can't be watched by calling ClipboardManager.getPrimaryClip or by adding
an onPrimaryClipChanged listener for notification when the clipboard
changes. This increases user privacy and disables malvertizing apps from
modifying the clipboard."

The next commit will introduce an alternative approach for the moment
we remove the listener from the SOClipboardHandler interface and it's
implementations.

<log>
Android NUI:

The SOClipboardHandler interface has been updated to remove the need
for a clipboard change listener.

SDK customers should update their implementations of this interface
as per the supplied ClipboardHandler.java.
</log>