Android NUI: Update Clipboard Usage Logic
This commit updates the way we choose whether to paste from the
internal or system clipboard now that a change listener cannot be used.

We now note the contents of the system clipboard after each cut/copy
operation is executed.

On a paste request, if the data leakage policy allows reading from the
system clipboard, we compare the system clipboard contents with our note
taken in any previous cut/copy operation. If they differ we paste from
the system clipboard. If not we paste from the internal clipboard.

If reading from of the system clipboard is prohibited we paste from
the internal clipboard.

<log>
Android NUI:

This commit resolves Bug 702614.

The text style can no longer be lost when on cut/copy operation.
</log>