Android NUI: Disable UI Buttons When Finishing The App
This commit addresses an issue, noted in Bug 702460, that the
toolbar buttons can be called when 'mDocView' is null.

The only circumstance I would see this being possible is when
we are in the process of exiting, but the buttons are still
available.

The solution presented here is to check the 'mFinished' flag in
the button handlers and exit, without taking any further action,
if set.

The bug report states the issue was not reproducible. I could
instigate something similar by opening a text document then
clicking the 'Back' button immediately followed by the "First Page"
button.

<log>
Bug 702460 - Application crashes when tap on 'First Page' button
             while open any .txt file.

is resolved.

Smart Office no longer crashes when UI buttons are clicked during
shutdown.
</log>