[Android] Force a save after signing a PDF document
In order to provide a more robust document signing process, we force a save after signing
like Adobe Acrobat.

This helps avoid some timing issues we have with document signatures when a document
is being saved in the background and the user resumes the app.

We also use the save completion handlers to identify when the save filename has
been selected by the user, and sign the document at that point just prior to saving.

<log>[Android] Force a save after signing a PDF document</log>