[Android MuPDF] Fix inconsistent return values for SeekableStream:read()
Our SeekeableStream implementation in MuPDFDoc is incorrect.
In SecureFS the read() call returns > 0 for a successful read, 0 for EOF
and negative for error codes.
SeekableStream::read() returns > 0 for successful read, negative for EOF and 0 for 'can't read'.
Errors will be thrown.

Fix up out implementation of the SeekableStream interface to match