Android - DocumentView.setScaleAndScroll should trigger ChangePageListener
Calls to setScaleAndScroll() don't result in ChangePageListener.onPage() being called.
In this commit we fix that by waiting for the result of setScaleAndScroll to
settle, and then calling the host view's setCurrentPage().

Also, in the reflow case, we need to reset the reflow width and height,
same as as done in onScaleEnd().

This solves the three issues reported by a customer
on Oct. 10 in Zen ticket 3227.

<log></log>