Android - improve initial page scaling for XLS
This commit fixes the algorithm for determining an initial scale factor
for a sheet, the first time it appears.

The algorithm computes the scale factor that would result in making
the average row height 0.15 inches.

The retail apps have a mechanism for saving the scale and scroll
values and reapplying them when a document is opened. This may
conflict with documents that were opened in a previous version.

<log></log>