Fix 703051 - crash after enabling invert content in dark mode and try to open files
The cause of the crash is trying to cast an ArDkBitmap to SOBitmap, but it's not
an SOBitmap. That's because in DocPageView we were doing 'new ArDkBitmap'.

In this commit we add ArDkBitmap.createBitmap(n,n,n,n), with implementations
for SOBitMap and MuPdfBitmap. Then in DocPageView, we call createBitmap
instead, which insures that we're getting the correct bitmap variant.

<log></log>