Fix 703055 Save button is disabled after cancelling Save As
This issue was likely introduced in
https://gitlab.artifex.com/smartoffice/sot/-/merge_requests/1733

The problem is that resetting the template status of the document
should not be done in the Cancel case.

The fix is to move resetting the template status into the success
case, and adding an empty cancel case for clarity.

This is done in two places, one of which appliers to no-=UI.

<log></log>