Android BBEdit: Reinstate The 'SaveTo' Button
Bug 703073 noted that the "SaveTo" button was missing from the
'File' toolbar.

The root cause of this was that the editor module flavour specific
resources, in this case the 'sodk_editor_custom_save_support' layout,
were not being picked up with the 'main' variant being used instead.

This commit addresses this issue:

- The use of the 'sodk_resources' project variable to hold the
  location of the editor module flavour specific resources has
  been replaced by setting them directly in the source set for
  particular flavour in the application build.gradle file.

- Duplicate BBEdit/BBEditEnt editor module dark mode color definitions
  have been removed from the application level colors.xml files.

<log>

Android BBEdit/BBD: Bug 703073 "'Save To' button is not displayed when open any file in BB edit" has been resolved.

The BBD and BBEdit apps now show the "SaveTo" button.
</log>