Fix 703102, Tapping at DONE does not move focus to the next tab idnex field
This was introduced when fixing bug #702754. That fix caused mEditingWidgetIndex
to be set to -1, causing this bug.

The fix here is to reset mEditingWidgetIndex and mEditingWidget only when
the editor being stopped is on a different page.

The other thing we do in this fix is to use IME_ACTION_NEXT instead of
IME_ACTION_DONE, which will give the keyboard button a different
appearance, that conveys "next".

<log></log>