Android - fix signature creation/positioning on 2nd page
This fixes relate to bugs #703097 and #703098
Creation and repositioning of signature blocks was not working
smoothly unless you were on page 1. This is fixed by correctlty
calculating the widget rectangle in view-space at key points in
the flow. The calculation is moved into its own function,
matchResizeRectToSig(), and called at those places.

While working on this I noticed that you can't tap on a signature
that appears over an edit field. There's a fix here to look for
tapped widgets from front to back.

<log></log>