Android - bring signatures in line with iOS 3.11
This commit brings Android in line with iOS in the following ways:

- Signing should not be possible in the Retail version. It should work
for Enterprise and Blackberry Dynamics (but not BBEdit).

- Signatures created during the current edit session can be
repositioned or deleted, but not those that were present when
the session began.

- the Dark Matter builds are set to enable form signing, since they
should be the same as Enterprise. Part of this change was to copy
the extra manifest file into src/darkmatterSigned.

<log></log>