Fix 703097, Crash when tapping at signature placeholder
This is a result of having DocMuPdfPageView hold a static value for
the page view that has the currently-edited form field. The value was
not getting reset when a new document is opened.

This fix is to move that reference to DocPdfView (not static),
and changing the corresponding usages. This way the value
is always null when a new document is opened.

<log></log>