Android - fix crashes with an invalid form field value during pause/resume
This change is meant to work around crashes that can happen when you
pause and resume at a time when a field has an invalid value. A validation
alert may or may not be displaying at the time you pause. I believe it's a side
effect of saving/reloading the document.

In this fix we cancel any outstanding form field edit before pausing fully.
This involves dismissing the alert and restoring the value that was present
when the edit began.

DocPageView and DocMuPdfPageView now have onPause(), which are
getting called at pause time. DocMuPdfPageView.oPause()
stopPreviousEditor(true), indicating that a cancel() should take place
on the form field, as opposed to a normal stop().

PDFFormTextEditor dismisses the alert and restores the original
value before proceeding with a normal stop().

When the file is reloaded upon coming to the foreground, the state of
the form editor, if it was active, is restored.

<log></log>