Android - fix signing for SO DK and MuPDF DK
This commit makes changes that allow digital signing of PDF files
to work properly. Three issues are addressed:

1.  DataLeakHandlers needs to use the onFilenameSelected() method
of the completion handler, in order to effect signing before the document
is actually saved. This is done in the retail apps, but needs to be
replicated in the sample apps as well.

2.  DocumentView.setDocDataLeakHandler should be called so DLH
gets used in the custom UI cases.

3.  Signing is executed on the worker thread.

<log></log>