Android - integrate Firebase/Crashlytics
As requested in bug:
https://bugs.ghostscript.com/show_bug.cgi?id=703145

Followed instructions from
https://firebase.google.com/docs/crashlytics/upgrade-sdk?platform=android&authuser=2

Summary:

- gradle and build tools versions are updated for smart-office-nui
(but not for the SDK sample apps)

- the changes just mentioned resulted in changes to postlink.py.
this is because the names and locations of files generated by gradle
in the 'outputs' directory is different.

- the file google-services.json is in app/src/release/regular, so it will
only be included in the Play Store version. I assume this is what we want.

- The optional NDK integration is not done.

- Proguard configurations for appstore, bbd and ent are changed,
to add '-keepattributes Signature', removing an error.

- When building sodk, I was getting a javadoc error about a
duplicate class. On the theory that the javadoc step for solib is
no longer erroneously generating an empty jar, I commented out the
'correction' for that in smartoffice-devkit-android-package.py,
and all seems well.

- changes to Proguard configurations were necessary, but Im not
sure why. Perhaps the R8 obfuscator behavior changed with the latest
build tools.

<log></log>