Fix 703158 - Form filling: Focused field goes off screen
I did not reproduce exactly this issue, but I saw another issue
where the view bounced to the right, then to the left,
finally showing the tapped-on field properly.

scrollCaretIntoView() is being called twice, and the first time
it's done with invalid text rectangles, those having not yet
been retrieved from the new form field.

In this fix we defer the first call to scrollCaretIntoView
until the text rects for the new field have been retrieved.

This seems to work well for fields that are tapped in,
and fields that are moved to through tabbing or the Next
key on the keyboard.

<log></log>