Android fix crash when tapping the page list
A crash can happen when you tap on a paeg in the page list
of a PDF file. This is because content pages and page list
pages are both instances of DocMuPdfPageView, and
tapping either will wind up calling dispatchTouchEvent().

In dispatchTouchEvent we do things related to form editing,
which shouldn't be done for page list pages.
So in this fix we only do them for pages in the
main content view.

<log></log>