Fix 703107 - Crash at second attempt to Print PDF
The issue here is that doing a print job causes a new Activity to appear,
triggering a background save right after PrintHeleprPdf.print()
has just done a save, with no intervening changes. This, combined
with mupdf incremental saving, is what's causing the problem.

In this fix, we do the save in DataLeakHandlers.printHandler(),
and set printing=true to prevent DataLeakHandlers.pauseHandler()
from also doing a save. Then we call PrintHeleprPdf.printPDF()
to print the file we saved. This way There's only one save done in
the lifecycle of the print request, and when it's done, we reload
the document.

In PrintHelperPDF, we added a new method, printPDF(), which prints
a PDF file given a path to the file, used by DataLeakHandlers.

Something similar needs to be done for the sample apps, mupdf-test
and solib-test. That's not covered by this fix.

<log></log>