Fix 703203 - Reproducible crash: libmupdf_java.so (fz_fill_text+164) (testDoc_not_signed.pdf)
This fix addresses the bug by doing several things.

1.  defer things that NUIDocView.onResume() does until we're sure that the previous
onPause() is finished. Pressing the Power button twice rapidly, as the tester does,
can lead to onResume being called while onPause is still underway. Here we wait
for onPause to be done. This fixes the so-called "blur" that the tester refers to.

2.  don't allow onPause to continue if a previous one is underway, allowing the
background file save to complete.

3.  Show a "please wait" dialog while onResume is waiting for pause to finish.
NUIDocView uses two different such dialogs, so that code is moved into a pair
of static Utilities functions.

4.  When pausing with an active form field, we'll clear the value while pausing.
It will be restored upon resume. This seemed to eliminate crashing in Javascript-
related code although I'm not entirely sure why.

5.  On document reloading, move the disable and re-enable of Javascript to the
worker thread.

6.  When jt is disabled, set the JS event listener to one that does nothing.
The mupdf Java API doesn't allow for setting it to null.

<log></log>