Android NUI: Fix Bug 701328 - When double tapped on an image inserted in the document “the keyboard pops up”, only the picture should be selected
Don't show the keyboard when/after double tap on an image object (which can be absolutely positioned) in pptx / ppt / docx / doc.
Don't select an image object again with SOSelectMode_DefaultUnit(Word selection) in pptx / ppt if the object is already selected when double tap on the object.
(Selecting the object with SOSelectMode_DefaultUnit can clear the current selection)

<log>Android NUI: Fix Bug 701328 - When double tapped on an image inserted in the document “the keyboard pops up”, only the picture should be selected</log>