Fix 702919, Font Cannot Be Changed
This was fixed on master at an earlier time, but got lopst when
NUIEditToolbar was introduced on Sept 23, 2020.

This has likely not worked since SOTextView and SOEditText were
reworked.
onClick(View v) is called with the TextView that’s wrapped inside
the SOTextView. This change checks for that case as well.

Also fix an issue where placing the cursor does not disable
the font controls. This was due to a missing SOTextView.setEnabled()

<log></log>