Fix 703222, Android NUI: Partial Blank Screen When Inserting Image
In NUIDocView, after the keyboard is raised or lowered, we wait a
beat (250 msec) for the layout to settle. This is how it was done
originally. A change was made in mid-2019 to use OnGlobalLayoutListener,
somewhat arbitrarily, which does seem to be causing this bug.
So now we go back to the original method.

<log></log>