Android NUI: Take Display Cutouts Into Account When Calculating Keyboard Height
Bug 703003 noted that, on some devices, placing the caret to show
the virtual keyboard caused the document view to be pushed off the
top of the screen.

The reason for this that the devices in question had a "cutout" area
at the top/bottom of the screen. The document view was being pushed under
the top opaque area by NUIDocView::onShowKeyboardPreventPush().

Display cutouts are available from Android 9 onwards. As of Android 11
the onShowKeyboardPreventPush() is not required. Thus this fix only
affects Android 9 and 10.

The solution is to, in portrait mode only, add the difference of the
top/bottom cutout heights to the keyboard height to correctly push
the view back down to the top of the visible screen.

An additional fudge factor is added to the cutout height calculation
when there is a bottom cutout. I can't explain why this is required,
but it does seem to be related to the hiding of the footer when
the keyboard is show. I expect the presence of a bottom cutout will
be relatively rare so this situation shouldn't be common.

The cutout heights are consumed and reset to 0 by the system, thus we
read and store them in onAttachedToWindow() before this happens.

<log>
Bug 703003: Menu tab disappears is resolved.

The top of the document view is no longer hidden when the keyboard
is displayed on devices configured with display cutouts.
</log>