Android NUI: Fix Crash Obtaining Uri Scheme
The following crash was noted on FireBase:

Attempt to invoke virtual method
'java.lang.String android.net.Uri.getScheme()' on a null object reference

    com.artifex.solib.FileUtils.getFileTypeExtension (FileUtils.java:696)
    com.artifex.sonui.editor.NUIView.makeNUIView (NUIView.java:89)
    com.artifex.sonui.editor.NUIView.start (NUIView.java:113)

The crash was due to attempting to dereference the getScheme()
function of a null Uri.

The Uri in getFileTypeExtension() can originate from the intent
getData() call, which can return null.

So we just handle the null case.

This will cause extensionFromUriFilename() to be called with a null Uri.
That case is handled also.

<log></log>