Android NUI: Fix Incorrect State After Dark Mode Toggle
It was noted that after toggling between day/night mode an
incorrect UI and document position could be shown.

  - Open a Word document from the Explorer view.
  - Switch to the native file manager and "Open In" another document
    type and move to the end of the document.
  - Toggle dark mode.
  - The UI color and options will now be for a Word document and the
    document will be on page 1.

Toggling dark mode causes NUIActivity::checkUIMode() to pause
the activity, to save the current state, before restarting it.

This causes DataLeakHandlers::pauseHandler() to save the document and
create an auto open entry. NUIDocView.currentNUIDocView is used to obtain
the current document state for this entry, however in this case it is
holding the NUIDocView for the previous document as that is not updated
until the activity is resumed.

The solution presented here updates NUIDocView.currentNUIDocView
when the activity is paused in addition to when it is resumed.

<log></log>