Android - Modify DocumentView to use Lifecycle APIs
Use Lifecycle APIs in DocumentView so it can handle pause,
resume, destroy, back button press, and configuration changes,
instead of relying on the Activity of which it's a part from needing
to do it.

This allows us to remove such handling from the SDK custom UI
sample activities, also done in this commit.

This requires any Activity that uses DocumentView to be a
FragmentActivity or AppCompatActivity, which I think is a
reasonable requirement.

<log></log>