Android NUI: Fix Crash On Null Uri
Firebase noted the following crash for v3.11.7:

Fatal Exception: java.lang.NullPointerException
Attempt to invoke virtual method 'java.lang.String android.net.Uri.getScheme()' on a null object reference

com.artifex.sonui.editor.NUIDocView$9.onGlobalLayout (NUIDocView.java:1388)

The issue was that NUIDocView was attempting to use the contents of a
null uri to obtain the path of the document to open.

A null uri is not expected at this point in proceedings.

We now handle this case by presenting an error dialog an exiting.

<log></log>