Android NUI: Prevent Leaving The App In A Bad State On Start Error
NUIDocView::onGlobalLayout() shows an error message and returns
when it can't obtain the path for the document to open.

This leaves the application in a state where the document view is
displayed, but the back button has no effect and pressing any tab
causes a crash.

To prevent this We now finish the activity in this case.

<log></log>